requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
psycopg2-binary==2.9.9
requests==2.31.0
urllib3==2.0.7
